# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Expense Approval API",
        "description": "API for handling expense approval system",
        "version": "1.0.0"
    },
    "paths": {
        "/api/approval-stages": {
            "post": {
                "tags": [
                    "Approval Stage"
                ],
                "summary": "Create a new approval stage",
                "description": "This endpoint creates a new approval stage in the approval workflow system.",
                "operationId": "52989e6a14178a727bcc848e3acf8044",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Approval stage created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Approval stage created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "approver_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input"
                    }
                }
            }
        },
        "/api/approval-stages/{id}": {
            "put": {
                "tags": [
                    "Approval Stage"
                ],
                "summary": "Update an existing approval stage",
                "description": "This endpoint updates an existing approval stage in the approval workflow system.",
                "operationId": "09e3ea8a8b87944defd865cd349851f9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the approval stage to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approval stage updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Approval stage updated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "approver_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input"
                    },
                    "404": {
                        "description": "Not found - the approval stage was not found"
                    }
                }
            }
        },
        "/api/approvers": {
            "post": {
                "tags": [
                    "Approver"
                ],
                "summary": "Create a new approver",
                "description": "This endpoint creates a new approver in the system.",
                "operationId": "8fddc9e935853028d340f48240c4f3ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Approver created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Approver created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input"
                    }
                }
            }
        },
        "/api/expense": {
            "post": {
                "tags": [
                    "Expense"
                ],
                "summary": "Create a new expense",
                "description": "This endpoint creates a new expense record in the system.",
                "operationId": "a8e3ae7ca07f1c973a526d413dc78a7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "integer",
                                        "example": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expense created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 1000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input"
                    }
                }
            }
        },
        "/api/expense/{id}/approve": {
            "patch": {
                "tags": [
                    "Expense"
                ],
                "summary": "Approve an expense",
                "description": "This endpoint approves an expense record in the system.",
                "operationId": "5ff2fd78df71a303e74d8a6c26c4dd0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense to approve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "approver_id"
                                ],
                                "properties": {
                                    "approver_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Expense approved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 1000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid input"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            }
        },
        "/api/expense/{id}": {
            "get": {
                "tags": [
                    "Expense"
                ],
                "summary": "Get a single expense record",
                "description": "This endpoint retrieves an expense record with its approval status.",
                "operationId": "051a76d052cc492fba7e7536cecfd1d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": 1000
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "approved"
                                        },
                                        "approvals": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "approver": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "approved"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Approval Stage",
            "description": "Approval Stage"
        },
        {
            "name": "Approver",
            "description": "Approver"
        },
        {
            "name": "Expense",
            "description": "Expense"
        }
    ]
}